+++ apps/openmw/mwgui/spellbuyingwindow.cpp
@@ -230,7 +230,7 @@ namespace MWGui
     {
         if (arg.button == SDL_CONTROLLER_BUTTON_A)
         {
-            if (mControllerFocus >= 0 && mControllerFocus < mSpellButtons.size())
+            if (mControllerFocus >= 0 && mControllerFocus < static_cast<int>(mSpellButtons.size()))
                 onSpellButtonClick(mSpellButtons[mControllerFocus].first);
         }
         else if (arg.button == SDL_CONTROLLER_BUTTON_B)
@@ -262,7 +262,7 @@ namespace MWGui
             mSpellButtons[mControllerFocus].first->setStateSelected(true);
         }
 
-        if (mControllerFocus >= 0 && mControllerFocus < mSpellButtons.size())
+        if (mControllerFocus >= 0 && mControllerFocus < static_cast<int>(mSpellButtons.size()))
         {
             // Scroll the list to keep the active item in view
             int line = mSpellButtons[mControllerFocus].second;
