+++ apps/openmw/mwgui/trainingwindow.cpp
@@ -251,7 +251,7 @@ namespace MWGui
     {
         if (arg.button == SDL_CONTROLLER_BUTTON_A)
         {
-            if (mControllerFocus >= 0 && mControllerFocus < mTrainingButtons.size())
+            if (mControllerFocus >= 0 && mControllerFocus < static_cast<int>(mTrainingButtons.size()))
                 onTrainingSelected(mTrainingButtons[mControllerFocus]);
         }
         else if (arg.button == SDL_CONTROLLER_BUTTON_B)
