+++ apps/openmw/mwgui/windowbase.cpp
@@ -27,7 +27,7 @@ int MWGui::wrap(int index, int max)
 
 void MWGui::setControllerFocus(std::vector<MyGUI::Button*> buttons, int index, bool focused)
 {
-    if (index >= 0 && index < buttons.size())
+    if (index >= 0 && index < static_cast<int>(buttons.size()))
         buttons[index]->setStateSelected(focused);
 }
